+++ lib/thirdparty/flutter_context_menu/core/models/context_menu.dart
@@ -36,7 +36,7 @@ class ContextMenu {
 
   /// A map of shortcuts to be bound to the context menu and the nested context menus.
   ///
-  /// Note: This overides the default shortcuts in [defaultMenuShortcuts] if any of the keys match.
+  /// Note: This overrides the default shortcuts in [defaultMenuShortcuts] if any of the keys match.
   Map<ShortcutActivator, VoidCallback> shortcuts;
 
   ContextMenu({