+++ lib/trios/settings/settings.mapper.dart
@@ -625,6 +625,23 @@ class SettingsMapper extends ClassMapperBase<Settings> {
         opt: true,
         def: DashboardModListSort.name,
       );
+  static bool _$dashboardUseLoadOrderForNameSort(Settings v) =>
+      v.dashboardUseLoadOrderForNameSort;
+  static const Field<Settings, bool> _f$dashboardUseLoadOrderForNameSort =
+      Field(
+        'dashboardUseLoadOrderForNameSort',
+        _$dashboardUseLoadOrderForNameSort,
+        opt: true,
+        def: false,
+      );
+  static bool _$modsPageUseLoadOrderForNameSort(Settings v) =>
+      v.modsPageUseLoadOrderForNameSort;
+  static const Field<Settings, bool> _f$modsPageUseLoadOrderForNameSort = Field(
+    'modsPageUseLoadOrderForNameSort',
+    _$modsPageUseLoadOrderForNameSort,
+    opt: true,
+    def: false,
+  );
   static bool _$checkIfGameIsRunning(Settings v) => v.checkIfGameIsRunning;
   static const Field<Settings, bool> _f$checkIfGameIsRunning = Field(
     'checkIfGameIsRunning',
@@ -737,6 +754,8 @@ class SettingsMapper extends ClassMapperBase<Settings> {
     #enableLauncherPrecheck: _f$enableLauncherPrecheck,
     #modUpdateBehavior: _f$modUpdateBehavior,
     #dashboardModListSort: _f$dashboardModListSort,
+    #dashboardUseLoadOrderForNameSort: _f$dashboardUseLoadOrderForNameSort,
+    #modsPageUseLoadOrderForNameSort: _f$modsPageUseLoadOrderForNameSort,
     #checkIfGameIsRunning: _f$checkIfGameIsRunning,
     #compressionLib: _f$compressionLib,
     #windowScaleFactor: _f$windowScaleFactor,
@@ -802,6 +821,12 @@ class SettingsMapper extends ClassMapperBase<Settings> {
       enableLauncherPrecheck: data.dec(_f$enableLauncherPrecheck),
       modUpdateBehavior: data.dec(_f$modUpdateBehavior),
       dashboardModListSort: data.dec(_f$dashboardModListSort),
+      dashboardUseLoadOrderForNameSort: data.dec(
+        _f$dashboardUseLoadOrderForNameSort,
+      ),
+      modsPageUseLoadOrderForNameSort: data.dec(
+        _f$modsPageUseLoadOrderForNameSort,
+      ),
       checkIfGameIsRunning: data.dec(_f$checkIfGameIsRunning),
       compressionLib: data.dec(_f$compressionLib),
       windowScaleFactor: data.dec(_f$windowScaleFactor),
@@ -935,6 +960,8 @@ abstract class SettingsCopyWith<$R, $In extends Settings, $Out>
     bool? enableLauncherPrecheck,
     ModUpdateBehavior? modUpdateBehavior,
     DashboardModListSort? dashboardModListSort,
+    bool? dashboardUseLoadOrderForNameSort,
+    bool? modsPageUseLoadOrderForNameSort,
     bool? checkIfGameIsRunning,
     CompressionLib? compressionLib,
     double? windowScaleFactor,
@@ -1034,6 +1061,8 @@ class _SettingsCopyWithImpl<$R, $Out>
     bool? enableLauncherPrecheck,
     ModUpdateBehavior? modUpdateBehavior,
     DashboardModListSort? dashboardModListSort,
+    bool? dashboardUseLoadOrderForNameSort,
+    bool? modsPageUseLoadOrderForNameSort,
     bool? checkIfGameIsRunning,
     CompressionLib? compressionLib,
     double? windowScaleFactor,
@@ -1110,6 +1139,10 @@ class _SettingsCopyWithImpl<$R, $Out>
       if (modUpdateBehavior != null) #modUpdateBehavior: modUpdateBehavior,
       if (dashboardModListSort != null)
         #dashboardModListSort: dashboardModListSort,
+      if (dashboardUseLoadOrderForNameSort != null)
+        #dashboardUseLoadOrderForNameSort: dashboardUseLoadOrderForNameSort,
+      if (modsPageUseLoadOrderForNameSort != null)
+        #modsPageUseLoadOrderForNameSort: modsPageUseLoadOrderForNameSort,
       if (checkIfGameIsRunning != null)
         #checkIfGameIsRunning: checkIfGameIsRunning,
       if (compressionLib != null) #compressionLib: compressionLib,
@@ -1248,6 +1281,14 @@ class _SettingsCopyWithImpl<$R, $Out>
     dashboardModListSort: data.get(
       #dashboardModListSort,
       or: $value.dashboardModListSort,
+    ),
+    dashboardUseLoadOrderForNameSort: data.get(
+      #dashboardUseLoadOrderForNameSort,
+      or: $value.dashboardUseLoadOrderForNameSort,
+    ),
+    modsPageUseLoadOrderForNameSort: data.get(
+      #modsPageUseLoadOrderForNameSort,
+      or: $value.modsPageUseLoadOrderForNameSort,
     ),
     checkIfGameIsRunning: data.get(
       #checkIfGameIsRunning,