+++ changelog.md
@@ -20,6 +20,9 @@
   - Option to have Favorited mods not pinned at the top of the mod list.
   - If you use Fast Rendering, its custom vmparams file (`fr.vmparams`) is now modified when you change RAM from within TriOS.
   - Option to show only disabled mods on the Dashboard mod list.
+  - Option when sorting mods by name to toggle between Starsector sorting (mod load order) and intuitive sorting.
+    - Starsector sorting: `   x`, `Z`, `a`.
+    - Intuitive sorting: `a`, `   x`, `Z`.
 - Fixed
   - Ship Viewer filters not working.
   - Custom launcher files are no longer hardcoded to use the game folder as their working directory.